+++ src/main/res/layout/interpolation.xml
@@ -10,6 +10,7 @@
     >
     <TextView
         android:id="@+id/start_title"
+        style="?attr/StaticText"
         android:layout_alignParentTop="true"
         android:layout_alignParentStart="true"
         android:layout_alignParentLeft="true"
@@ -92,6 +93,7 @@
     </RelativeLayout>
     <TextView
         android:id="@+id/end_title"
+        style="?attr/StaticText"
         android:layout_alignParentTop="true"
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
