+++ src/main/res/layout/simple_text_row.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- inner classes must be specified using <view class="classname" /> due to dollar sign -->
+<view class="de.blau.android.views.SimpleTextRow" xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="horizontal">
+    <LinearLayout 
+			android:orientation="horizontal"
+			android:baselineAligned="false"
+			android:layout_width="fill_parent"
+			android:layout_height="wrap_content"
+			style="?attr/Layout"
+			android:saveEnabled="false"
+            android:gravity="center_vertical">
+	    <TextView
+	        android:id="@+id/textKey"
+	        style="?attr/StaticText"
+	        android:saveEnabled="false"
+	        android:layout_width="fill_parent"
+	        android:layout_height="wrap_content"
+	        android:layout_marginBottom="1dp"
+	        android:layout_marginRight="1dp"
+            android:layout_marginEnd="1dp"
+	        android:layout_weight="3"
+	        android:dropDownWidth="wrap_content"
+	        android:minHeight="34dp"
+	        android:paddingLeft="5dp"
+	        android:paddingRight="5dp"
+	        android:maxLines="3" />	    
+	    <de.blau.android.views.CustomAutoCompleteTextView
+	        android:id="@+id/textValue"
+	        android:saveEnabled="false"
+	        android:layout_width="fill_parent"
+	        android:layout_height="wrap_content"
+	        android:layout_marginLeft="0dp"
+            android:layout_marginStart="0dp"
+	        android:layout_marginBottom="1dp"
+	        android:layout_weight="2"
+	        android:completionThreshold="1"
+	        android:dropDownWidth="wrap_content"	        
+	        android:hint="@string/value"
+	        android:inputType="textNoSuggestions"
+	        android:minHeight="34dp"
+	        android:paddingLeft="5dp"
+	        android:paddingRight="5dp"
+			android:maxLines="1" />
+	</LinearLayout>
+</view>
